made all entity with overriding equals and hash code added validation

+++ src/main/java/com/example/PrisonManagement/Service/InfirmaryService.java
@@ -7,6 +7,6 @@ public interface InfirmaryService {
     List<Infirmary> getAllInfirmaries();
     Optional<Infirmary> getPrisonerFromInfirmaryById(Integer id);
     Infirmary createInfirmary(Infirmary infirmary);
-    Infirmary updateInfirmary(Long id, Infirmary infirmary);
-    void deleteInfirmary(Long id);
+    Infirmary updateInfirmary(Integer id, Infirmary infirmary);
+    void deleteInfirmary(Integer id);
 }
